requirements.txt
https://github.com/Rapptz/discord.py/archive/rewrite.zip#egg=discord.py
py-trello